Fix slack signing body decoding
When I originally built this I had worked with a `slack-sansio`
diff to perform the decoding inside of `validate_request_signature`
but when I moved over to writing the tests here I switched to the
upstream `slack-sansio` without the change and wrote the tests for
that incorrect behavior. This fixes the issue entirely within sirbot.